+++ src/com/reason/ide/folding/DuneFoldingBuilder.java
@@ -21,8 +21,8 @@ public class DuneFoldingBuilder extends FoldingBuilderEx {
         PsiTreeUtil.processElements(root, element -> {
             if (isMultiline(element.getTextRange(), document)) {
                 FoldingDescriptor fold = null;
-                if (element instanceof RPsiStanza) {
-                    fold = foldStanza((RPsiStanza) element);
+                if (element instanceof RPsiDuneStanza) {
+                    fold = foldStanza((RPsiDuneStanza) element);
                 } else if (DuneTypes.INSTANCE.C_SEXPR == element.getNode().getElementType()) {
                     fold = fold(element);
                 }
@@ -42,7 +42,7 @@ public class DuneFoldingBuilder extends FoldingBuilderEx {
                 < document.getLineNumber(range.getEndOffset());
     }
 
-    private @Nullable FoldingDescriptor foldStanza(@NotNull RPsiStanza root) {
+    private @Nullable FoldingDescriptor foldStanza(@NotNull RPsiDuneStanza root) {
         RPsiDuneFields fields = ORUtil.findImmediateFirstChildOfClass(root, RPsiDuneFields.class);
         return fields == null ? null : new FoldingDescriptor(root, fields.getTextRange());
     }