+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 (_Tags are copied from [babel](https://github.com/babel/babel/blob/master/CHANGELOG.md)_)
 
 ## Unreleased
-
+          
+- :nail_care: Display dune-project fields in structure view
 - :bug: [#358](https://github.com/giraud/reasonml-idea-plugin/issues/358) Incorrect resolution with module in pattern
 - :bug: [#323](https://github.com/giraud/reasonml-idea-plugin/issues/323) Method declarations in .ml files should link to their implementations
 - :house: ![o] Parse ternary if in OCaml